Added to core / acq testing
